Improve error reporting of recovery test 027_stream_regress
Previously, the 027_stream_regress test reported the full contents of
regression.diffs upon a test failure, when the standby and the primary
were still alive.  If a test fails quite badly, the amount of
information reported can be really high, bloating the reports in the
buildfarm, the CI, or even local runs.

In most cases, we have noticed that having all this information is not
necessary when attempting to identify the source of a problem in this
test.  This commit changes the situation by including the head and tail
of regression.diffs in the reports generated on failure rather than its
full contents, building upon e0b78ce85b50 to optionally control the size
of the reports with the new environment variable
PG_TEST_FILE_READ_LINES.

This will perhaps require some more tuning, but the hope is to reduce
some of the buildfarm report bloat while making the information good
enough to deduce what is happening when something is going wrong, be it
in the buildfarm or some tests run in the CI, at least.

Suggested-by: Andres Freund <andres@anarazel.de>
Author: Nazir Bilal Yavuz <byavuz81@gmail.com>
Reviewed-by: Michael Paquier <michael@paquier.xyz>
Discussion: https://postgr.es/m/CAN55FZ1D6KXvjSs7YGsDeadqCxNF3UUhjRAfforzzP0k-cE=bA@mail.gmail.com